+++ tests/test_mc_helper.cpp
@@ -16,7 +16,8 @@ TYPED_TEST(McHelper, CumulativeResult0WithOne)
 	std::vector<hep::mc_result<T>> one_result = {
 		hep::mc_result<T>(1000, T(10.0), T(20.0))
 	};
-	auto result = hep::cumulative_result0(one_result.begin(), one_result.end());
+	auto result = hep::accumulate<hep::weighted_with_variance>(
+		one_result.begin(), one_result.end());
 
 	EXPECT_EQ( one_result.front().calls() , result.calls() );
 	EXPECT_NEAR( one_result.front().value() , result.value() , T(1e-10) );
@@ -30,7 +31,8 @@ TYPED_TEST(McHelper, CumulativeResult1WithOne)
 	std::vector<hep::mc_result<T>> one_result = {
 		hep::mc_result<T>(1000, T(10.0), T(20.0))
 	};
-	auto result = hep::cumulative_result1(one_result.begin(), one_result.end());
+	auto result = hep::accumulate<hep::weighted_equally>(one_result.begin(),
+		one_result.end());
 
 	EXPECT_EQ( one_result.front().calls() , result.calls() );
 	EXPECT_NEAR( one_result.front().value() , result.value() , T(1e-10) );
@@ -46,8 +48,8 @@ TYPED_TEST(McHelper, CumulativeResult0WithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	auto result = hep::cumulative_result0(two_results.begin(),
-		two_results.end());
+	auto result = hep::accumulate<hep::weighted_with_variance>(
+		two_results.begin(), two_results.end());
 
 	// should give two times the calls
 	EXPECT_EQ( 2 * two_results.front().calls() , result.calls() );
@@ -67,7 +69,7 @@ TYPED_TEST(McHelper, CumulativeResult1WithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	auto result = hep::cumulative_result1(two_results.begin(),
+	auto result = hep::accumulate<hep::weighted_equally>(two_results.begin(),
 		two_results.end());
 
 	// should give two times the calls
@@ -88,7 +90,7 @@ TYPED_TEST(McHelper, CumulativeResult1WithThree)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 	};
-	auto result = hep::cumulative_result1(two_results.begin(),
+	auto result = hep::accumulate<hep::weighted_equally>(two_results.begin(),
 		two_results.end());
 
 	// should give three times the calls
@@ -104,7 +106,8 @@ TYPED_TEST(McHelper, ChiSquareDof0WithZero)
 	typedef TypeParam T;
 
 	std::vector<hep::mc_result<T>> zero_results;
-	T result = hep::chi_square_dof0(zero_results.begin(), zero_results.end());
+	T result = hep::chi_square_dof<hep::weighted_with_variance>(
+		zero_results.begin(), zero_results.end());
 
 	EXPECT_NEAR( T() , result , T(1e-10) );
 }
@@ -114,7 +117,8 @@ TYPED_TEST(McHelper, ChiSquareDof1WithZero)
 	typedef TypeParam T;
 
 	std::vector<hep::mc_result<T>> zero_results;
-	T result = hep::chi_square_dof1(zero_results.begin(), zero_results.end());
+	T result = hep::chi_square_dof<hep::weighted_equally>(zero_results.begin(),
+		zero_results.end());
 
 	EXPECT_NEAR( T() , result , T(1e-10) );
 }
@@ -126,7 +130,8 @@ TYPED_TEST(McHelper, ChiSquareDof0WithOne)
 	std::vector<hep::mc_result<T>> one_result = {
 		hep::mc_result<T>(1000, T(10.0), T(20.0))
 	};
-	T result = hep::chi_square_dof0(one_result.begin(), one_result.end());
+	T result = hep::chi_square_dof<hep::weighted_with_variance>(
+		one_result.begin(), one_result.end());
 
 	EXPECT_EQ( std::numeric_limits<T>::infinity() , result );
 }
@@ -138,7 +143,8 @@ TYPED_TEST(McHelper, ChiSquareDof1WithOne)
 	std::vector<hep::mc_result<T>> one_result = {
 		hep::mc_result<T>(1000, T(10.0), T(20.0))
 	};
-	T result = hep::chi_square_dof1(one_result.begin(), one_result.end());
+	T result = hep::chi_square_dof<hep::weighted_equally>(one_result.begin(),
+		one_result.end());
 
 	EXPECT_EQ( std::numeric_limits<T>::infinity() , result );
 }
@@ -152,7 +158,8 @@ TYPED_TEST(McHelper, ChiSquareDof0WithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	T result = hep::chi_square_dof0(two_results.begin(), two_results.end());
+	T result = hep::chi_square_dof<hep::weighted_with_variance>(
+		two_results.begin(), two_results.end());
 
 	EXPECT_NEAR( T() , result , T(1e-10) );
 }
@@ -166,7 +173,8 @@ TYPED_TEST(McHelper, ChiSquareDof1WithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	T result = hep::chi_square_dof1(two_results.begin(), two_results.end());
+	T result = hep::chi_square_dof<hep::weighted_equally>(two_results.begin(),
+		two_results.end());
 
 	EXPECT_NEAR( T() , result , T(1e-10) );
 }